+++ apps/vscode/src/main.ts
@@ -43,6 +43,8 @@ import { activateDiagram } from "./providers/diagram/diagram";
 import { activateCodeFormatting } from "./providers/format";
 import { activateOptionEnterProvider } from "./providers/option";
 import { activateBackgroundHighlighter } from "./providers/background";
+import { activateYamlLinks } from "./providers/yaml-links";
+import { activateYamlFilepathCompletions } from "./providers/yaml-filepath-completions";
 import { activateContextKeySetter } from "./providers/context-keys";
 import { CommandManager } from "./core/command";
 import { createQuartoExtensionApi, QuartoExtensionApi } from "./api";
@@ -204,6 +206,12 @@ export async function activate(context: vscode.ExtensionContext): Promise<Quarto
 
   // background highlighter
   activateBackgroundHighlighter(context, engine);
+
+  // yaml document links
+  activateYamlLinks(context);
+
+  // yaml filepath completions
+  activateYamlFilepathCompletions(context);
 
   // context setter
   activateContextKeySetter(context, engine);